add rest api

+++ src/LearnIt.MAUI/Services/LocalCategoryService.cs
@@ -18,10 +18,10 @@ namespace LearnIt.MAUI.Services
             return _localStorageService.Get(LocalStorageConstants.IS_SKIP_START_PAGE_KEY, false);
         }
 
-        public void InsertCategoryIds(int[] newIds)
+        public void InsertCategoryIds(long[] newIds)
         {
             string idsFromStorage = _localStorageService.Get(LocalStorageConstants.SELECTED_CATEGORIES_KEY, "");
-            var oldIds = JsonSerializeExtension.TryDeserializeArray<int>(idsFromStorage);
+            var oldIds = JsonSerializeExtension.TryDeserializeArray<long>(idsFromStorage);
             var ids = newIds.Union(oldIds).OrderBy(x => x);
             var json = JsonSerializer.Serialize(ids);
             _localStorageService.Add(LocalStorageConstants.SELECTED_CATEGORIES_KEY, json);
